+++ R/googleAuthR_auth.R
@@ -5,10 +5,8 @@
 Authentication <- R6::R6Class(
   "Authentication",
   public = list(
-    websites = "initial",
     token = NULL,
-    shiny = FALSE,
-    app_url = NULL
+    method = NULL
   ),
   lock_objects = F,
   parent_env = emptyenv()
@@ -70,17 +68,24 @@ Authentication <- R6::R6Class(
 gar_auth <- function(token = NULL,
                      new_user = FALSE) {
   
+  httr_file <- getOption("googleAuthR.httr_oauth_cache")
+  if(inherits(httr_file, "logical")){
+    httr_file <- ".httr-oauth"
+  }
+  
   if(new_user) {
     Authentication$set("public", "token", NULL, overwrite=TRUE)
-    Authentication$set("public", "websites", data.frame(siteURL="None", permissionLevel="N/A"), overwrite=TRUE)
-    if(file.exists(".httr-oauth")){
-      myMessage("Removing old credentials ...", level=2)
-      file.remove(".httr-oauth")     
+    if(file.exists(httr_file)){
+      myMessage("Removing old credentials ... ", httr_file, level=2)
+      file.remove(httr_file)     
     }
     
   }
   
   if(is.null(token)) {
+    
+    ## check if current cached token has same scopes as those set in options
+    check_cached_scopes()
     
     endpoint <- httr::oauth_endpoints("google")
     
@@ -93,36 +98,80 @@ gar_auth <- function(token = NULL,
                                          scope = getOption("googleAuthR.scopes.selected"), 
                                          cache = getOption("googleAuthR.httr_oauth_cache"))   
 
-    
-    
     stopifnot(is_legit_token(google_token))
     
     Authentication$set("public", "token", google_token, overwrite=TRUE)
+    Authentication$set("public", "method", "new_token", overwrite=TRUE)
     
   } else {
-    
+    ## supplied a file path or Token object
     if(is_legit_token(token)) {
+      ## a token object, just return it back
       google_token <- token
+      Authentication$set("public", "method", "passed_token", overwrite=TRUE)
     } else {
-      myMessage("Reading token from file path", level = 2)
-      google_token <- try(suppressWarnings(readRDS(token)[[1]]), silent = TRUE)
-      
-      if(is.error(google_token)) {
-        stop(sprintf("Cannot read token from alleged .rds file:\n%s",
-                          token))
-      } else if(!is_legit_token(google_token)) {
-          
-        stop(sprintf("File does not contain a proper token:\n%s", token))
-
-      }
+      ## a file path to a token
+      options("googleAuthR.httr_oauth_cache" = token)
+      google_token <- read_cache_token(token_path = token)
+      google_token$cache_path <- token
+      Authentication$set("public", "method", "filepath", overwrite=TRUE)
     }
     
     Authentication$set("public", "token", google_token, overwrite=TRUE)
     
   }
-  
+  gar_token_info()
   return(invisible(Authentication$public_fields$token)) 
   
+}
+
+#' Get current token summary
+#' 
+#' Get details on the current active auth token to help debug issues
+#' 
+#' @export
+gar_token_info <- function(){
+  token <- Authentication$public_fields$token
+  method <- Authentication$public_fields$method
+  
+  myMessage("Token cache file: ", token$cache_path, level = 3)
+  myMessage("Scopes: ", paste(token$params$scope, collapse = " "), level = 2)
+  myMessage("Hash: ", token$hash(), level = 2)
+  
+  if(!is.null(token$app$key)){
+    myMessage("App key: ", token$app$key, level = 2)
+  }
+
+  myMessage("Method: ", method, level = 2)
+  
+  ## service
+  if(!is.null(token$secrets)){
+    myMessage("Type: ", token$secrets$type, level = 2)
+    myMessage("ProjectID: ", token$secrets$project_id, level = 2)
+    myMessage("Client email: ", token$secrets$client_email, level = 2)
+    myMessage("ClientID: ", token$secrets$client_id, level = 2)
+  }
+}
+
+## httr cache files such as .httr-oauth can hold multiple tokens for different scopes.
+## this selects the token that covers the scope given
+read_cache_token <- function(token_path = getOption("googleAuthR.httr_oauth_cache")){
+  if(inherits(token_path, "logical")){
+    token_path <- ".httr-oauth"
+  }
+  myMessage("Reading token from file path", level = 2)
+  google_token <- try(suppressWarnings(readRDS(token_path)), silent = TRUE)
+  if(is.error(google_token)) {
+    stop(sprintf("Cannot read token from alleged .rds file:\n%s",
+                 token_path))
+  }
+  
+  if(inherits(google_token, "list")){
+    myMessage("Found a list, return token in first element.")
+    google_token <- google_token[[1]]
+  }
+  
+  google_token
 }
 
 #' Retrieve Google token from environment and configs for httr
@@ -145,7 +194,7 @@ get_google_token <- function(shiny_return_token=NULL) {
     
     
   } else { #shiny session
-    
+    Authentication$set("public", "method", "shiny", overwrite=TRUE)
     token <- shiny_return_token
     
   }
@@ -178,7 +227,7 @@ token_exists <- function() {
     if(file.exists(httr_cache)) {
       myMessage(paste("NOTE: a ", httr_cache ,
                       " file exists in current working",
-                      "directory.\n Run gar_auth() to use the",
+                      "directory.\n Run authentication function to use the",
                       "credentials cached for this session."), level=3)
     } else {
       myMessage(paste("No ", httr_cache ,
@@ -206,6 +255,12 @@ is_legit_token <- function(x) {
   
   if(!inherits(x, "Token2.0")) {
     myMessage("Not a Token2.0 object. Found:", class(x), level=2)
+    if(!inherits(x, "list")){
+      if(inherits(x[[1]], "Token2.0")){
+        myMessage("Its a list of Token2.0 objects though")
+        return(TRUE)
+      }
+    }
     return(FALSE)
   }
   
@@ -267,16 +322,113 @@ gar_auth_service <- function(json_file, scope = getOption("googleAuthR.scopes.se
   scope <- paste(scope, collapse=" ")
   
   if(is.null(secrets$private_key)){
-    stop("private_key not found in JSON - have you downloaded the correct JSON file? (Service Account Keys, not service account client)")
+    stop("$private_key not found in JSON - have you downloaded the correct JSON file? 
+         (Service Account Keys, not service account client)")
   }
   
   google_token <- httr::oauth_service_token(endpoint, secrets, scope)
   
   Authentication$set("public", "token", google_token, overwrite=TRUE)
-  
+  Authentication$set("public", "method", "service_json", overwrite=TRUE)
   myMessage("Returning service token", level=1)
   
   return(invisible(Authentication$public_fields$token))
   
 }
 
+#' Authenticate on Google Compute Engine
+#' 
+#' This takes the metadata auth token in a Google Compute Engine instance as authentication source
+#' 
+#' @param service_account Specify a different service account from the \code{default}
+#' @inheritParams gar_shiny_getToken
+#' 
+#' @details 
+#' 
+#' \code{service_account} is \code{default} or the service account email 
+#' e.g. \code{"service-account-key-json@projectname.iam.gserviceaccount.com"}
+#' 
+#' Google Compute Engine instances come with their own authentication tokens.  
+#' 
+#' It has no refresh token so you need to call for a fresh token after approx. one hour. 
+#' The metadata token will refresh itself when it has about 60 seconds left.
+#' 
+#' You can only use for scopes specified when creating the instance.
+#' 
+#' If you want to use them make sure their service account email is added to accounts you want to get data from.
+#' 
+#' If this function is called on a non-Google Compute Engine instance it will return \code{NULL}
+#' 
+#' @return A token 
+#' @export
+#' @family authentication functions
+gar_gce_auth <- function(service_account = "default",
+                         client.id     = getOption("googleAuthR.webapp.client_id"),
+                         client.secret = getOption("googleAuthR.webapp.client_secret")){
+  
+  call_url <- sprintf("http://metadata.google.internal/computeMetadata/v1/instance/service-accounts/%s/token",
+                      service_account)
+  
+  ## will be an error if not on GCE
+  req <- try(httr::GET(call_url, httr::add_headers("Metadata-Flavor" = "Google")), silent = TRUE)
+  
+  if(is.error(req)){
+    myMessage("Not detected as being on Google Compute Engine", level = 2)
+    return(NULL)
+  }
+
+  token <- httr::content(req, type = "application/json")
+  
+  gar_app <- httr::oauth_app("google", key = client.id, secret = client.secret)
+  
+  scope_list <- getOption("googleAuthR.scope")
+  
+  token_formatted <-
+    httr::Token2.0$new(app = gar_app,
+                       endpoint = httr::oauth_endpoints("google"),
+                       credentials = list(access_token = token$access_token,
+                                          token_type = token$token_type,
+                                          expires_in = token$expires_in,
+                                          refresh_token = NULL),
+                       params = list(scope = scope_list, type = NULL,
+                                     use_oob = FALSE, as_header = TRUE),
+                       cache_path = getOption("googleAuthR.httr_oauth_cache"))
+  
+  myMessage("Authenticated on Google Compute Engine", level = 2)
+  
+  ## for other google auth on a server (such as Google Analytics) need to manually do tokens via OOB
+  options(httr_oob_default = TRUE)
+  Authentication$set("public", "method", "gce_auth", overwrite=TRUE)
+  
+  ## puts it in environment
+  gar_auth(token_formatted)
+
+}
+
+
+#' Check token scopes
+## check if scopes are set correctly
+check_cached_scopes <- function(){
+  
+  httr_cache <- getOption("googleAuthR.httr_oauth_cache")
+  if(class(httr_cache) == "logical"){
+    httr_cache <- ".httr-oauth"
+  }
+  google_token <- try(suppressWarnings(readRDS(httr_cache)[[1]]), silent = TRUE)
+
+  out <- TRUE
+  if(!is.error(google_token)){
+    if(!all(google_token$params$scope %in%
+            getOption("googleAuthR.scopes.selected"))){
+      warning(paste0("option(googleAuthR.scopes.selected) not same scopes as current cached token ", httr_cache, ", will need reauthentication.  
+                     \nToken scopes: ", 
+                     paste(google_token$params$scope, collapse = " "),
+                     "\ngetOption(googleAuthR.scopes.selected): ",
+                     paste(getOption("googleAuthR.scopes.selected"), collapse = " "),
+                     collapse = " "), call. = FALSE)
+      out <- FALSE
+    }
+  }
+  
+  out
+}