+++ tests/testthat/test-gaAuthR.R
@@ -61,6 +61,7 @@ test_that("Can authenticate .httr passed as a file", {
 
 test_that("Can authenticate .httr looking for existing file", {
   skip_on_cran()
+  options(googleAuthR.httr_oauth_cache = "httr-oauth.rds")
   expect_s3_class(gar_auth(), "Token2.0")
   
 })