+++ NEWS.md
@@ -1,6 +1,20 @@
-# News Updates
+# v0.5.0
+
+## Major changes
+
+* Correct bug of incorrect redirect URL when no port (#45)
+* Catch error for retry if no status response at all
+* Fix bug where unnamed customConfigs were ignored
+* Fixes to discovery API package creations, order of parameters and NULL parameters
+* Add authentication option when using Google Compute Engine `gar_gce_auth()` (#52)
+* Add a warning if the cached `.httr-oauth` token has different scopes to the ones specified at time of authentication (#53)
+* Add debug body aid: if `option(googleAuthR.verbose = 0)` then a request with a body will write to a file `request_debug.rds` in working directory
+* Passing in a file location token to `gar_auth("file-location.rds")` will only load the first element `[[1]]` if that token is a list of `Token2.0` class objects
+* Add debug tool `gar_token_info()` which will report on current authentication. Available at `options(googleAuthR.verbose = 2)`
 
 # v0.4.0
+
+## Major changes
 
 * Add client based authentication in JavaScript plus example app
 * Add check to `gar_auth_service` to see if you have downloaded right JSON file
@@ -18,13 +32,17 @@
 * `gar_auth` respects renamed `.httr-oauth` tokens now via `getOption("googleAuthR.httr_oauth_cache")`
 * Add link to Github repo with auto-generated packages: `https://github.com/MarkEdmondson1234/autoGoogleAPI`
 
-# v0.3.1 - CRAN
+# v0.3.1
+
+## Major changes
 
 * Add link to [example shiny app](https://mark.shinyapps.io/googleAuthRexample/)
 * Add `option(googleAuthR.rawResponse)` - skip API checks on response - should now work
 * A successfull request is now classed as all response codes matching ^20 e.g. 201, 204 etc.
 
 # v0.3.0 
+
+## Major changes
 
 * Document default options in `?googleAuthR`
 * Add `option(googleAuthR.rawResponse)` - skip API checks on response.
@@ -40,6 +58,8 @@
 * Migrated shiny functions to Shiny Modules (#27)
 
 # v0.2 
+
+## Major changes
 
 * Added ability to add your own custom headers to requests via `customConfig` in `gar_api_generator`
 * Add 'localhost' to shiny URL detection. 
@@ -58,6 +78,8 @@
 * added gzip to headers
 
 # v0.1 - CRAN
+
+## Major changes
 
 * Shiny compatibility
 * Local authentication compatibility